- Introduce DEFAULT_START_TIME_GPS_NS, and DEFAULT_STOP_TIME_GPS_NS macros. - Change all error message macros to GST_ERROR_OBJECT(). - Be a little more careful with the offset and timestamp of the output   buffers:  the read_series() can clip the requested interval so the buffer   should be set from the properties of the time series returned from this   function, not from the parameters passed to the function. - Check that the blocksize property is an integer multiple of the sample   size.

+++ src/plugins/gstlal_framesrc.h
@@ -61,7 +61,7 @@ typedef struct {
 
 	LIGOTimeGPS start_time;
 	LIGOTimeGPS stop_time;
-	int next_sample;
+	guint64 next_sample;
 
 	FrStream *stream;
 	LALTYPECODE series_type;
